[CL2-6567] extracted EventCard and refactored it (WIP)
