+++ front/app/components/EventCard/messages.ts
@@ -1,0 +1,12 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  startsAt: {
+    id: 'app.components.EventCard.startsAt',
+    defaultMessage: 'Starts at',
+  },
+  endsAt: {
+    id: 'app.components.EventCard.endsAt',
+    defaultMessage: 'Ends at',
+  },
+});
