+++ front/app/translations/ar-SA.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "المرحلة الحالية",
   "app.containers.Projects.editProject": "تعديل المشروع",
   "app.containers.Projects.endedOn": "انتهى في {date}",
-  "app.containers.Projects.endsAt": "ينتهي عند",
+  "app.components.EventCard.endsAt": "ينتهي عند",
   "app.containers.Projects.events": "الأحداث",
   "app.containers.Projects.goBackToList": "عودة إلى القائمة",
   "app.containers.Projects.header": "المشاريع",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "مشاركة",
   "app.containers.Projects.shareThisProject": "مشاركة هذا المشروع",
   "app.containers.Projects.spentBudget": "الميزانية المعيّنة",
-  "app.containers.Projects.startsAt": "يبدأ عند",
+  "app.components.EventCard.startsAt": "يبدأ عند",
   "app.containers.Projects.submitMyExpenses": "أرسل نفقاتي",
   "app.containers.Projects.survey": "الاستبيان",
   "app.containers.Projects.takeThePoll": "شارك في الاستطلاع",
