+++ front/app/translations/da-DK.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Igangværende fase",
   "app.containers.Projects.editProject": "Rediger dette projekt",
   "app.containers.Projects.endedOn": "Sluttede den {date}",
-  "app.containers.Projects.endsAt": "Slutter den",
+  "app.components.EventCard.endsAt": "Slutter den",
   "app.containers.Projects.events": "Begivenheder",
   "app.containers.Projects.goBackToList": "Gå tilbage til listen",
   "app.containers.Projects.header": "{tenantName, select, DeloitteDK {Verdensmål} other {Projekter}}",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Del",
   "app.containers.Projects.shareThisProject": "Del dette projekt",
   "app.containers.Projects.spentBudget": "Tildelt budget ",
-  "app.containers.Projects.startsAt": "Starter den",
+  "app.components.EventCard.startsAt": "Starter den",
   "app.containers.Projects.submitMyExpenses": "Indsend mine udgifter",
   "app.containers.Projects.survey": "Spørgeundersøgelse",
   "app.containers.Projects.takeThePoll": "Tag afstemningen",
