+++ front/app/translations/de-DE.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Aktuelle Phase",
   "app.containers.Projects.editProject": "Projekt bearbeiten",
   "app.containers.Projects.endedOn": "Endete am {date}",
-  "app.containers.Projects.endsAt": "Endet um",
+  "app.components.EventCard.endsAt": "Endet um",
   "app.containers.Projects.events": "Veranstaltungen",
   "app.containers.Projects.goBackToList": "Zurück zur Liste",
   "app.containers.Projects.header": "Projekte",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Teilen",
   "app.containers.Projects.shareThisProject": "Das Projekt teilen",
   "app.containers.Projects.spentBudget": "Zugewiesenes Budget",
-  "app.containers.Projects.startsAt": "Beginnt um",
+  "app.components.EventCard.startsAt": "Beginnt um",
   "app.containers.Projects.submitMyExpenses": "Meine Ausgaben einreichen",
   "app.containers.Projects.survey": "Umfrage",
   "app.containers.Projects.takeThePoll": "Nehmen Sie an der Abstimmung teil",
