+++ front/app/translations/es-CL.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Fase actual",
   "app.containers.Projects.editProject": "Editar este proyecto",
   "app.containers.Projects.endedOn": "Terminó el {date}",
-  "app.containers.Projects.endsAt": "Termina a las",
+  "app.components.EventCard.endsAt": "Termina a las",
   "app.containers.Projects.events": "Eventos",
   "app.containers.Projects.goBackToList": "Volver a la lista",
   "app.containers.Projects.header": "Proyectos",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Compartir",
   "app.containers.Projects.shareThisProject": "Compartir este proyecto",
   "app.containers.Projects.spentBudget": "Presupuesto asignado",
-  "app.containers.Projects.startsAt": "Empieza a las",
+  "app.components.EventCard.startsAt": "Empieza a las",
   "app.containers.Projects.submitMyExpenses": "Presentar mis gastos",
   "app.containers.Projects.survey": "Consulta",
   "app.containers.Projects.takeThePoll": "Responder la consulta ciudadana",
