+++ front/app/translations/es-ES.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Fase actual",
   "app.containers.Projects.editProject": "Editar este proyecto",
   "app.containers.Projects.endedOn": "Terminó el {date}",
-  "app.containers.Projects.endsAt": "Termina en",
+  "app.components.EventCard.endsAt": "Termina en",
   "app.containers.Projects.events": "Eventos",
   "app.containers.Projects.goBackToList": "Volver a la lista",
   "app.containers.Projects.header": "Proyectos",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Compartir",
   "app.containers.Projects.shareThisProject": "Compartir este proyecto",
   "app.containers.Projects.spentBudget": "Presupuesto asignado",
-  "app.containers.Projects.startsAt": "Comienza en",
+  "app.components.EventCard.startsAt": "Comienza en",
   "app.containers.Projects.submitMyExpenses": "Presentar mis gastos",
   "app.containers.Projects.survey": "Encuesta",
   "app.containers.Projects.takeThePoll": "Haz la consulta ciudadana",
