+++ front/app/translations/fr-FR.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Phase en cours",
   "app.containers.Projects.editProject": "Modifier ce projet",
   "app.containers.Projects.endedOn": "Terminé le {date}",
-  "app.containers.Projects.endsAt": "Termine le",
+  "app.components.EventCard.endsAt": "Termine le",
   "app.containers.Projects.events": "Evénements",
   "app.containers.Projects.goBackToList": "Retour à la liste",
   "app.containers.Projects.header": "{tenantName, select, frw {Communes} other {Projets}}",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Partager",
   "app.containers.Projects.shareThisProject": "Partager ce projet",
   "app.containers.Projects.spentBudget": "Budget alloué",
-  "app.containers.Projects.startsAt": "Commence le",
+  "app.components.EventCard.startsAt": "Commence le",
   "app.containers.Projects.submitMyExpenses": "Soumettre mes dépenses",
   "app.containers.Projects.survey": "Enquête",
   "app.containers.Projects.takeThePoll": "{tenantName, select, vitrysurseine {Répondre à l'enquête} other {Répondre au sondage}}",
