+++ front/app/translations/kl-GL.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Igangværende fase",
   "app.containers.Projects.editProject": "Rediger dette projekt",
   "app.containers.Projects.endedOn": "Ulloq {date} taamaatippoq",
-  "app.containers.Projects.endsAt": "Ends at",
+  "app.components.EventCard.endsAt": "Ends at",
   "app.containers.Projects.events": "Pisimasut",
   "app.containers.Projects.goBackToList": "Allattuiffimmut uterit",
   "app.containers.Projects.header": "Suliniutit ",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Del",
   "app.containers.Projects.shareThisProject": "Del dette projekt",
   "app.containers.Projects.spentBudget": "Missingersuutit tunniunneqarsimasut",
-  "app.containers.Projects.startsAt": "Starts at",
+  "app.components.EventCard.startsAt": "Starts at",
   "app.containers.Projects.submitMyExpenses": "Akiligassakka nassiutikkit",
   "app.containers.Projects.survey": "Immersuilluni apersuineq",
   "app.containers.Projects.takeThePoll": "Tag afstemningen",
