+++ front/app/translations/mi.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Current phase",
   "app.containers.Projects.editProject": "Edit project",
   "app.containers.Projects.endedOn": "Ended on {date}",
-  "app.containers.Projects.endsAt": "Ends at",
+  "app.components.EventCard.endsAt": "Ends at",
   "app.containers.Projects.events": "Events",
   "app.containers.Projects.goBackToList": "Go back to the list",
   "app.containers.Projects.header": "Kaupapa",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Share",
   "app.containers.Projects.shareThisProject": "Share this project",
   "app.containers.Projects.spentBudget": "Assigned budget",
-  "app.containers.Projects.startsAt": "Starts at",
+  "app.components.EventCard.startsAt": "Starts at",
   "app.containers.Projects.submitMyExpenses": "Submit my expenses",
   "app.containers.Projects.survey": "Survey",
   "app.containers.Projects.takeThePoll": "Take the poll",
