+++ front/app/translations/nb-NO.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Nåværende fase",
   "app.containers.Projects.editProject": "Rediger prosjekt",
   "app.containers.Projects.endedOn": "Ble avsluttet {date}",
-  "app.containers.Projects.endsAt": "Slutter den",
+  "app.components.EventCard.endsAt": "Slutter den",
   "app.containers.Projects.events": "Hendelser",
   "app.containers.Projects.goBackToList": "Gå tilbake til listen",
   "app.containers.Projects.header": "Prosjekter",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Del",
   "app.containers.Projects.shareThisProject": "Del dette prosjekt",
   "app.containers.Projects.spentBudget": "Tildelt budsjett",
-  "app.containers.Projects.startsAt": "Starter den",
+  "app.components.EventCard.startsAt": "Starter den",
   "app.containers.Projects.submitMyExpenses": "Send inn mine utgifter",
   "app.containers.Projects.survey": "Spørreundersøkelse",
   "app.containers.Projects.takeThePoll": "Bli med på avstemmingen",
