+++ front/app/translations/nl-BE.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Huidige fase",
   "app.containers.Projects.editProject": "Bewerk dit project",
   "app.containers.Projects.endedOn": "Beëindigd op {date}",
-  "app.containers.Projects.endsAt": "Eindigt op",
+  "app.components.EventCard.endsAt": "Eindigt op",
   "app.containers.Projects.events": "Activiteiten",
   "app.containers.Projects.goBackToList": "Ga terug naar de lijst",
   "app.containers.Projects.header": "Projecten",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Delen",
   "app.containers.Projects.shareThisProject": "Deel dit project",
   "app.containers.Projects.spentBudget": "Uitgegeven bedrag",
-  "app.containers.Projects.startsAt": "Begint op",
+  "app.components.EventCard.startsAt": "Begint op",
   "app.containers.Projects.submitMyExpenses": "Bevestig mijn uitgaven",
   "app.containers.Projects.survey": "Enquête",
   "app.containers.Projects.takeThePoll": "Neem deel aan de peiling",
