+++ front/app/translations/pl-PL.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Aktualny etap",
   "app.containers.Projects.editProject": "Edytuj projekt",
   "app.containers.Projects.endedOn": "Zakończone {date}",
-  "app.containers.Projects.endsAt": "Kończy się o",
+  "app.components.EventCard.endsAt": "Kończy się o",
   "app.containers.Projects.events": "Wydarzenia",
   "app.containers.Projects.goBackToList": "Wróć do listy",
   "app.containers.Projects.header": "Projekty",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Podziel się",
   "app.containers.Projects.shareThisProject": "Podziel się tym projektem",
   "app.containers.Projects.spentBudget": "Przydzielony budżet",
-  "app.containers.Projects.startsAt": "Zaczyna się o",
+  "app.components.EventCard.startsAt": "Zaczyna się o",
   "app.containers.Projects.submitMyExpenses": "Prześlij moje wydatki",
   "app.containers.Projects.survey": "Ankieta",
   "app.containers.Projects.takeThePoll": "Weź udział w sondażu",
