+++ front/app/translations/pt-BR.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Fase atual",
   "app.containers.Projects.editProject": "Editar projeto",
   "app.containers.Projects.endedOn": "Terminado em {date}",
-  "app.containers.Projects.endsAt": "Termina em",
+  "app.components.EventCard.endsAt": "Termina em",
   "app.containers.Projects.events": "Eventos",
   "app.containers.Projects.goBackToList": "Voltar para listagem",
   "app.containers.Projects.header": "Projetos",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Compartilhar",
   "app.containers.Projects.shareThisProject": "Compartilhe este projeto",
   "app.containers.Projects.spentBudget": "Orçamento atribuído",
-  "app.containers.Projects.startsAt": "Inicia em",
+  "app.components.EventCard.startsAt": "Inicia em",
   "app.containers.Projects.submitMyExpenses": "Enviar as minhas despesas",
   "app.containers.Projects.survey": "Pesquisa",
   "app.containers.Projects.takeThePoll": "Faça a sondagem",
