+++ front/app/translations/ro-RO.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Faza curentă",
   "app.containers.Projects.editProject": "Editați acest proiect",
   "app.containers.Projects.endedOn": "Încheiat la {data}",
-  "app.containers.Projects.endsAt": "Se termină la",
+  "app.components.EventCard.endsAt": "Se termină la",
   "app.containers.Projects.events": "Evenimente",
   "app.containers.Projects.goBackToList": "Du-te înapoi la listă",
   "app.containers.Projects.header": "Proiecte",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Distribuie",
   "app.containers.Projects.shareThisProject": "Distribuie acest proiect",
   "app.containers.Projects.spentBudget": "Bugetul alocat",
-  "app.containers.Projects.startsAt": "Începe la",
+  "app.components.EventCard.startsAt": "Începe la",
   "app.containers.Projects.submitMyExpenses": "Trimiteți-mi cheltuielile",
   "app.containers.Projects.survey": "Chestionar",
   "app.containers.Projects.takeThePoll": "Completează sondajul de tip poll",
