+++ front/app/translations/sr-CS.json
@@ -2255,7 +2255,7 @@
   "app.containers.Projects.currentPhase": "Trenutna faza",
   "app.containers.Projects.editProject": "Uredite projekat",
   "app.containers.Projects.endedOn": "Završeno {date}",
-  "app.containers.Projects.endsAt": "Završava se u",
+  "app.components.EventCard.endsAt": "Završava se u",
   "app.containers.Projects.events": "Događaji",
   "app.containers.Projects.goBackToList": "Vratite se na listu",
   "app.containers.Projects.header": "Projekti",
@@ -2296,7 +2296,7 @@
   "app.containers.Projects.share": "Podeli",
   "app.containers.Projects.shareThisProject": "Podelite ovaj projekat",
   "app.containers.Projects.spentBudget": "Dodeljeni budžet",
-  "app.containers.Projects.startsAt": "Počinje u",
+  "app.components.EventCard.startsAt": "Počinje u",
   "app.containers.Projects.submitMyExpenses": "Predajte moje troškove",
   "app.containers.Projects.survey": "Upitnik",
   "app.containers.Projects.takeThePoll": "Popunite anketu",
